// src/app/generated/vm-api/model/powerState.ts
/*
Copyright 2021 Carnegie Mellon University. All Rights Reserved. 
 Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.
*/

/**
 * Player VM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type PowerState = 'Unknown' | 'On' | 'Off' | 'Suspended';

export const PowerState = {
    Unknown: 'Unknown' as PowerState,
    On: 'On' as PowerState,
    Off: 'Off' as PowerState,
    Suspended: 'Suspended' as PowerState
};